Extract constants, utilities, core world logic into own modules
